pyproject.toml
[tool.poetry]
name = "todolist3"
version = "0.1.0"
description = ""
authors = ["Павел <chernovpp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.6"
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.2.0"
django-filter = "^23.1"
drf-spectacular = "^0.26.2"
pydantic = "^1.10.7"
requests = "^2.30.0"
pytelegrambotapi = "^4.11.0"

[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.1"
pre-commit = "^3.2.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = 'todolist.settings'
python_files = [
    'tests.py',
    'test_*.py',
    '*_tests.py'
]

[tool.black]
line-length = 120
target-version = ['py310']
skip-string-normalization = true
skip-magic-trailing-coma = true
exclude = '''
(
  /(
      \.git
    | \.github
    | \.venv
    | \.pytest_cache
    | deploy
  )/
)
'''
